Inject context in xorm session
* Objective

This commit objective is about generating a new xorm session that has
the request context injected.

* Why

There are many use-cases for having the request.context injected, but we
need this specifically to make the sentry transaction to reach the xorm
before/after middleware.

* How

This commit makes use of xorm.Engine.Context() to create a new session
with an injected context